Make ref contextual menus localisable
Looking at these I started wondering whether it’d be a good idea
to remove the inactive items in there. I can see the point of them
making things more consistent across menus – but the idea of a
contextual menu started off focusing on the commands that are actually
usable in the current context.

Besides, the menus are already quite long.